requirements.txt
google-cloud-storage>=2.10.0
google-auth>=2.22.0
google-cloud-bigquery==3.14.1
pandas==2.2.1
python-dotenv>=1.0.0
requests>=2.31.0
db-dtypes==1.2.0
hubspot-api-client==8.1.0
tqdm==4.66.2 